adding new method find, to WineController

+++ database/migrations/2025_05_26_140831_create_wines_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('winery');
             $table->string('variety');
-            $table->tinyInteger('vintage');
+            $table->integer('vintage');
             $table->string('country');
             $table->timestamps();
         });
